proj/requirements.txt
Django==2.1.3
django-webpack-loader==0.6.0
pytz==2018.7
djangorestframework==3.9.4
psycopg2-binary==2.8.3
gevent